AK3: Use the optional OSS msm_drm module as appropriate
https://github.com/cupid-development/android_device_xiaomi_marble/commit/eee64379280d5bc680e91371679d788b63fe5039

Fixed the issue with automatic brightness not working on the ROMs based on latest OSS kernel.

+++ anykernel.sh
@@ -504,7 +504,12 @@ if ${is_hyperos_fw}; then
 		fi
 	fi
 	if ${use_oss_msm_drm}; then
-		cp -f ${home}/_alt/OSS-msm_drm.ko ${home}/_vendor_dlkm_modules/msm_drm.ko
+		if [ -f /vendor/etc/displayconfig/display_id_4630946370515662721.xml ] || [ -f /vendor/etc/displayconfig/display_id_4630946480857061761.xml ]; then
+			# https://github.com/cupid-development/android_device_xiaomi_marble/commit/eee64379280d5bc680e91371679d788b63fe5039
+			cp -f ${home}/_alt/OSS-msm_drm-2.ko ${home}/_vendor_dlkm_modules/msm_drm.ko
+		else
+			cp -f ${home}/_alt/OSS-msm_drm.ko ${home}/_vendor_dlkm_modules/msm_drm.ko
+		fi
 	fi
 	unset use_oss_msm_drm skip_option_oss_msm_drm
 fi
